feat: 메일 전송 로직을 mail 모듈로 분리 및 적용 (#15)
