+++ jobterview-api/build.gradle.kts
@@ -4,11 +4,8 @@ plugins {
 
 dependencies {
     implementation(project(":jobterview-domain"))
+    implementation(project(":jobterview-mail"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-mail")
     implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0")
-
-    implementation("org.springframework.boot:spring-boot-starter-thymeleaf")
-    implementation("nz.net.ultraq.thymeleaf:thymeleaf-layout-dialect")
 }
