+++ jobterview-api/src/main/kotlin/jobterview/api/ServerApplication.kt
@@ -1,11 +1,14 @@
 package jobterview.api
 
+import jobterview.batch.config.MailAsyncConfig
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
+import org.springframework.context.annotation.Import
 
 @SpringBootApplication(
-    scanBasePackages = ["jobterview.api", "jobterview.domain"],
+    scanBasePackages = ["jobterview.api", "jobterview.domain", "jobterview.mail", "jobterview.common"],
 )
+@Import(MailAsyncConfig::class)
 class ServerApplication
 
 fun main(args: Array<String>) {
