--- jobterview-api/src/main/kotlin/jobterview/api/mail/template/MailTemplate.kt
@@ -1,6 +1,0 @@
-package jobterview.api.mail.template
-
-interface MailTemplate {
-
-    fun render(data: Map<String, Any>): String
-}
