+++ jobterview-batch/build.gradle.kts
@@ -4,9 +4,8 @@ plugins {
 
 dependencies {
     implementation(project(":jobterview-domain"))
+    implementation(project(":jobterview-mail"))
 
     implementation("org.springframework.boot:spring-boot-starter-batch")
-    implementation("org.springframework.boot:spring-boot-starter-mail")
-
     testImplementation("org.springframework.batch:spring-batch-test")
 }
