+++ jobterview-common/src/main/kotlin/jobterview/common/config/AsyncConfig.kt
@@ -1,0 +1,8 @@
+package jobterview.common.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.annotation.EnableAsync
+
+@EnableAsync
+@Configuration
+class AsyncConfig
