+++ jobterview-mail/src/main/kotlin/jobterview/mail/config/MailAsyncConfig.kt
@@ -3,13 +3,11 @@ package jobterview.batch.config
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.scheduling.annotation.AsyncConfigurer
-import org.springframework.scheduling.annotation.EnableAsync
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
 import java.util.concurrent.Executor
 
-@EnableAsync
 @Configuration
-class AsyncConfig : AsyncConfigurer {
+class MailAsyncConfig : AsyncConfigurer {
 
     @Bean("mailExecutor")
     override fun getAsyncExecutor(): Executor {
