+++ jobterview-mail/src/main/kotlin/jobterview/mail/template/MailTemplate.kt
@@ -1,0 +1,6 @@
+package jobterview.mail.template
+
+interface MailTemplate {
+
+    fun render(data: Map<String, Any>): String
+}
