+++ settings.gradle.kts
@@ -8,3 +8,4 @@ include("jobterview-common")
 include("jobterview-domain")
 include("jobterview-batch")
 include("jobterview-api")
+include("jobterview-mail")
